Added basic user registration function

+++ base/views.py
@@ -1,11 +1,13 @@
 # This is my core views file
 
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-
 from django.db.models import Q
+from django.contrib.auth.forms import UserCreationForm
 from base.models import Room, Topic
 from base.forms import RoomForm
 
@@ -22,9 +24,13 @@ from base.forms import RoomForm
 
 #logic for login page
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('Homepage')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -41,13 +47,33 @@ def loginPage(request):
             messages.error(request, 'Username or Password does not exist')
 
 
-    context ={}
+    context ={'page':page}
     return render(request, 'base/login_register.html',context)
 
 #logic for user logout
 def logoutUser(request):
     logout(request)
     return redirect('Homepage')
+
+#logic for registration
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('Homepage')
+
+        else:
+            messages.error(request, "An error occured during registration")
+            
+    return render(request, 'base/login_register.html',{'form':form})
+
+
 
 #Home page logic
 def home(request):
@@ -64,6 +90,8 @@ def home(request):
 
     context = {'rooms':rooms, 'topics': topics, 'room_count':room_count}
     return  render(request, 'base/home.html',context)
+
+
 
 
 #Room room page logic
@@ -84,6 +112,7 @@ def chats(request):
 
 
 # Logic for creating new room
+@login_required(login_url ='/login') # Restricts this function to authenicated users only
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -94,11 +123,17 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+
 #Logic for updating room
+@login_required(login_url ='/login') # Restricts this function to authenicated users only
 def updateRoom(request, pk):
     room = Room.objects.get(id =pk)
     room = Room.objects.get(id=pk)
     form = RoomForm(instance = room)
+
+    #Prevents non-super admins from using this function
+    if request.user != room.host:
+        return HttpResponse('You have no permission to do this')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance =room)
@@ -110,9 +145,14 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html',context)
     
 #logic for deleting rooms
-
+@login_required(login_url ='/login') # Restricts this function to authenicated users only
 def deleteRoom(request,pk):
     room =Room.objects.get(id=pk)
+
+        #Prevents non-super admins from using this function
+    if request.user != room.host:
+        return HttpResponse('You have no permission to do this')
+    
     if request.method == 'POST':
         room.delete()
         return redirect("Homepage")
